+++ code/src/cljs/sixsq/nuvla/ui/utils/semantic_ui_extensions.cljs
@@ -231,26 +231,7 @@
   [{:keys [button-text on-confirm danger-msg header content trigger open on-close]}]
   (let [tr                   (subscribe [::i18n-subs/tr])
         confirmed?           (r/atom (nil? danger-msg))
-        clicked?             (r/atom false)
-        checkbox-warning-msg (fn [danger-msg]
-                               ;; force confirmed to false at open of modal
-                               (reset! confirmed? false)
-                               (fn [danger-msg]
-                                 [ui/Checkbox {:label     danger-msg
-                                               :checked   @confirmed?
-                                               :fitted    true
-                                               :on-change #(swap! confirmed? not)}]))
-
-        button-confirm       (fn []
-                               (reset! clicked? false)
-                               (fn []
-                                 [Button {:text     (str/capitalize button-text)
-                                          :negative true
-                                          :disabled (or (not @confirmed?) @clicked?)
-                                          :loading  @clicked?
-                                          :active   true
-                                          :on-click #(do (reset! clicked? true)
-                                                         (on-confirm))}]))]
+        clicked?             (r/atom false)]
     (fn [{:keys [button-text on-confirm danger-msg header content trigger open on-close]}]
       [ui/Modal (cond->
                   {:on-click   (fn [event]
@@ -260,7 +241,9 @@
                    :trigger    trigger
                    :on-close   (fn [& args]
                                  (when on-close
-                                   (apply on-close args)))}
+                                   (apply on-close args))
+                                 (reset! confirmed? (nil? danger-msg))
+                                 (reset! clicked? false))}
                   (some? open) (assoc :open open))
 
        (when header [ui/ModalHeader header])
@@ -270,10 +253,19 @@
         (when danger-msg
           [ui/Message {:error true}
            [ui/MessageHeader {:style {:margin-bottom 10}} (@tr [:danger-action-cannot-be-undone])]
-           [ui/MessageContent [checkbox-warning-msg danger-msg]]])]
+           [ui/MessageContent [ui/Checkbox {:label     danger-msg
+                                            :checked   @confirmed?
+                                            :fitted    true
+                                            :on-change #(swap! confirmed? not)}]]])]
 
        [ui/ModalActions
-        [button-confirm]]])))
+        [Button {:text     (str/capitalize button-text)
+                 :negative true
+                 :disabled (or (not @confirmed?) @clicked?)
+                 :loading  @clicked?
+                 :active   true
+                 :on-click #(do (reset! clicked? true)
+                                (on-confirm))}]]])))
 
 
 (defn TimeAgo